Move metadata builder to System.Reflection.Metadata.Ecma335 namespace
