+++ src/Compilers/CSharp/Portable/CodeGen/EmitArrayInitializer.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Linq;
+using System.Reflection.Metadata.Ecma335;
 using Microsoft.CodeAnalysis.CodeGen;
 using Microsoft.CodeAnalysis.CSharp.Symbols;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -310,14 +311,14 @@ namespace Microsoft.CodeAnalysis.CSharp.CodeGen
             // the initial size is a guess.
             // there is no point to be precise here as MemoryStream always has N + 1 storage 
             // and will need to be trimmed regardless
-            var writer = new Cci.BlobBuilder(initializers.Length * 4);
+            var writer = new BlobBuilder(initializers.Length * 4);
 
             SerializeArrayRecursive(writer, initializers);
 
             return writer.ToImmutableArray();
         }
 
-        private void SerializeArrayRecursive(Cci.BlobBuilder bw, ImmutableArray<BoundExpression> inits)
+        private void SerializeArrayRecursive(BlobBuilder bw, ImmutableArray<BoundExpression> inits)
         {
             if (inits.Length != 0)
             {
