+++ src/Compilers/Core/CodeAnalysisTest/PEWriter/BlobUtilitiesTests.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using Microsoft.Cci;
+using System.Reflection.Metadata.Ecma335;
 using Xunit;
 
 namespace Microsoft.CodeAnalysis.UnitTests.PEWriter
