+++ src/Compilers/Core/Portable/CodeGen/ILBuilderEmit.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
+using System.Reflection.Metadata.Ecma335;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CodeGen
@@ -704,7 +705,7 @@ namespace Microsoft.CodeAnalysis.CodeGen
             this.GetCurrentWriter().WriteInt64(int64);
         }
 
-        private static void WriteOpCode(Cci.BlobBuilder writer, ILOpCode code)
+        private static void WriteOpCode(BlobBuilder writer, ILOpCode code)
         {
             var size = code.Size();
             if (size == 1)
@@ -724,7 +725,7 @@ namespace Microsoft.CodeAnalysis.CodeGen
             }
         }
 
-        private Cci.BlobBuilder GetCurrentWriter()
+        private BlobBuilder GetCurrentWriter()
         {
             return this.GetCurrentBlock().Writer;
         }
