+++ src/Compilers/Core/Portable/Compilation/Compilation.cs
@@ -10,6 +10,7 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Reflection.Metadata;
+using System.Reflection.Metadata.Ecma335;
 using System.Reflection.PortableExecutable;
 using System.Text;
 using System.Threading;
@@ -1998,7 +1999,7 @@ namespace Microsoft.CodeAnalysis
                         changes,
                         cancellationToken);
 
-                    Cci.MetadataSizes metadataSizes;
+                    MetadataSizes metadataSizes;
                     writer.WriteMetadataAndIL(pdbWriter, metadataStream, ilStream, out metadataSizes);
                     writer.GetMethodTokens(updatedMethods);
