+++ src/Compilers/Core/Portable/PEWriter/Constants.cs
@@ -25,15 +25,6 @@ namespace Microsoft.Cci
         public const SignatureTypeCode SignatureTypeCode_ValueType = (SignatureTypeCode)0x11;
     }
 
-    internal enum HeapSizeFlag : byte
-    {
-        StringHeapLarge = 0x01, // 4 byte uint indexes used for string heap offsets
-        GuidHeapLarge = 0x02,   // 4 byte uint indexes used for GUID heap offsets
-        BlobHeapLarge = 0x04,   // 4 byte uint indexes used for Blob heap offsets
-        EnCDeltas = 0x20,       // Indicates only EnC Deltas are present
-        DeletedMarks = 0x80,    // Indicates metadata might contain items marked deleted
-    }
-
     internal static class TokenTypeIds
     {
         internal const int Module = 0x00000000;
