+++ src/Compilers/Core/Portable/PEWriter/ManagedResource.cs
@@ -7,6 +7,7 @@ using System.IO;
 using Microsoft.CodeAnalysis;
 using Roslyn.Utilities;
 using System.Globalization;
+using System.Reflection.Metadata.Ecma335;
 
 namespace Microsoft.Cci
 {
