+++ src/Compilers/Core/Portable/System.Reflection.Metadata/Ecma335/Blob.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
-
-namespace Microsoft.Cci
+namespace System.Reflection.Metadata.Ecma335
 {
     internal struct Blob
     {
