+++ src/Compilers/Core/Portable/System.Reflection.Metadata/Ecma335/BlobBuilder.Enumerators.cs
@@ -1,11 +1,10 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
 using System.Diagnostics;
 using System.Collections;
 using System.Collections.Generic;
 
-namespace Microsoft.Cci
+namespace System.Reflection.Metadata.Ecma335
 {
     partial class BlobBuilder
     {
