+++ src/Compilers/Core/Portable/System.Reflection.Metadata/Ecma335/EncFuncCode.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+namespace System.Reflection.Metadata
+{
+    internal enum EncFuncCode
+    {
+        Default = 0,
+        AddMethod = 1,
+        AddField = 2,
+        AddParameter = 3,
+        AddProperty = 4,
+        AddEvent = 5
+    }
+}
