+++ src/Compilers/Core/Portable/System.Reflection.Metadata/Ecma335/MetadataSerializer.cs
@@ -1,12 +1,10 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Linq;
-using System.Reflection.Metadata.Ecma335;
 
-namespace Microsoft.Cci
+namespace System.Reflection.Metadata.Ecma335
 {
     internal sealed class StandaloneDebugMetadataSerializer : MetadataSerializer
     {
