+++ src/Compilers/Core/Portable/System.Reflection.Metadata/PortableExecutable/CorHeader.cs
