+++ src/Compilers/Core/Portable/System.Reflection.Metadata/PortableExecutable/DirectoryEntry.cs
