+++ src/Compilers/Core/Portable/System.Reflection.Metadata/PortableExecutable/NtHeader.cs
