+++ src/Compilers/Core/Portable/System.Reflection.Metadata/PortableExecutable/SectionHeader.cs
